+++ scripts/prepare_sector_network.py
@@ -58,7 +58,9 @@ spatial = SimpleNamespace()
 logger = logging.getLogger(__name__)
 
 
-def define_spatial(nodes, options, offshore_buses_fn=None, buses_h2_file=None):
+def define_spatial(
+    nodes, options, offshore_buses_fn=None, buses_h2_file=None, tyndp_scenario=None
+):
     """
     Namespace for spatial.
 
@@ -207,17 +209,37 @@ def define_spatial(nodes, options, offshore_buses_fn=None, buses_h2_file=None):
     if options["h2_topology_tyndp"] and buses_h2_file:
         spatial.h2_tyndp = SimpleNamespace()
         buses_h2 = gpd.read_file(buses_h2_file).set_index("bus_id")
-        spatial.h2_tyndp.nodes = pd.Index(
-            (buses_h2.index + " Z1").append(buses_h2.index + " Z2")
-        )
-        spatial.h2_tyndp.locations = pd.Index(np.tile(buses_h2.index + " Z2", 2))
-        spatial.h2_tyndp.country = pd.Index(np.tile(buses_h2.country, 2))
-        spatial.h2_tyndp.x = pd.Index(np.tile(buses_h2.x, 2))
-        spatial.h2_tyndp.y = pd.Index(np.tile(buses_h2.y, 2))
-        spatial.h2_tyndp.df = pd.DataFrame(
-            vars(spatial.h2_tyndp),
-            index=pd.Index((buses_h2.index + " Z1").append(buses_h2.index + " Z2")),
-        )
+
+        if options["h2_zones_tyndp"]:
+            spatial.h2_tyndp.nodes = pd.Index(
+                (buses_h2.index + " Z1").append(buses_h2.index + " Z2")
+            )
+            spatial.h2_tyndp.locations = pd.Index(np.tile(buses_h2.index + " Z2", 2))
+            spatial.h2_tyndp.country = pd.Index(np.tile(buses_h2.country, 2))
+            spatial.h2_tyndp.x = pd.Index(np.tile(buses_h2.x, 2))
+            spatial.h2_tyndp.y = pd.Index(np.tile(buses_h2.y, 2))
+            spatial.h2_tyndp.df = pd.DataFrame(
+                vars(spatial.h2_tyndp),
+                index=pd.Index((buses_h2.index + " Z1").append(buses_h2.index + " Z2")),
+            )
+            spatial.buses_h2_z1 = spatial.h2_tyndp.nodes[
+                ~spatial.h2_tyndp.nodes.str.contains("IB")
+                & spatial.h2_tyndp.nodes.str.contains("Z1")
+            ]
+            spatial.buses_h2_z2 = spatial.h2_tyndp.nodes[
+                ~spatial.h2_tyndp.nodes.str.contains("IB")
+                & spatial.h2_tyndp.nodes.str.contains("Z2")
+            ]
+        else:
+            spatial.h2_tyndp.nodes = buses_h2.index
+            spatial.h2_tyndp.locations = buses_h2.index
+            spatial.h2_tyndp.country = buses_h2.country
+            spatial.h2_tyndp.x = buses_h2.x
+            spatial.h2_tyndp.y = buses_h2.y
+            spatial.buses_h2_z1 = pd.Index([])
+            spatial.buses_h2_z2 = spatial.h2_tyndp.nodes[
+                ~spatial.h2_tyndp.nodes.str.contains("IB")
+            ]
 
     # methanol
 
@@ -488,7 +510,7 @@ def create_network_topology(
     return topo
 
 
-def create_h2_topology_tyndp(n, fn_h2_network):
+def create_h2_topology_tyndp(n, fn_h2_network, options):
     """
     Create a TYNDP H2 network topology from the TYNDP H2 reference grid.
 
@@ -498,16 +520,20 @@ def create_h2_topology_tyndp(n, fn_h2_network):
         Network to create H2 topology for
     fn_h2_network : str
         Pointing to the input TYNDP H2 reference grid csv file
+    options : dict
+        Dictionary of configuration options. Key options include:
+        - h2_zones_tyndp : bool
 
     Returns
     -------
     pd.DataFrame with columns bus0, bus1, length, underwater_fraction
     """
 
+    suffix = "H2 Z2" if options["h2_zones_tyndp"] else "H2"
     # load H2 pipes
     h2_pipes = pd.read_csv(fn_h2_network, index_col=0)
     h2_pipes = h2_pipes.assign(
-        bus0=h2_pipes.bus0 + " H2 Z2", bus1=h2_pipes.bus1 + " H2 Z2"
+        bus0=h2_pipes.bus0 + f" {suffix}", bus1=h2_pipes.bus1 + f" {suffix}"
     )
     h2_pipes["length"] = h2_pipes.apply(haversine, axis=1, args=(n,))
 
@@ -2338,9 +2364,9 @@ def add_electricity_grid_connection(n, costs):
     ]
 
 
-def add_h2_production_tyndp(n, nodes, buses_h2_z1, costs, options={}):
+def add_h2_production_tyndp(n, nodes, buses_h2, costs, options={}):
     """
-    Add TYNDP electrolysers for Z1 and Z2, and optionally add SMR, SMR CC and ATR.
+    Add TYNDP electrolyzers for Z1 and Z2, and optionally add SMR, SMR CC and ATR.
 
     Parameters
     ----------
@@ -2348,8 +2374,8 @@ def add_h2_production_tyndp(n, nodes, buses_h2_z1, costs, options={}):
         The PyPSA network container object
     nodes : pd.Index
         Pandas Index of electricity node locations/nodes
-    buses_h2_z1 : SimpleNamespace
-        Namespace object with spatial nodes of H2 Z1 buses
+    buses_h2 : pd.Index
+        Pandas Index of hydrogen nodes to which H2 production technologies will connect
     costs : pd.DataFrame
         Technology cost assumptions
     options : dict, optional
@@ -2367,16 +2393,14 @@ def add_h2_production_tyndp(n, nodes, buses_h2_z1, costs, options={}):
         The function modifies the network object in-place by adding components.
     """
 
-    logger.info("Adding Z1 and Z2 dummy electrolysers.")
+    suffix = "H2 Z1" if options["h2_zones_tyndp"] else "H2"
+
+    # Add electrolysis to Z1 or in NT case general H2 nodes
     n.add(
         "Link",
-        (nodes.index + " H2 Z1 Electrolysis").append(
-            nodes.index + " H2 Z2 Electrolysis"
-        ),
-        bus0=np.tile(nodes.index, 2),
-        bus1=np.append(
-            nodes.country.values + " H2 Z1", (nodes.country.values + " H2 Z2")
-        ),
+        nodes.index + f" {suffix} Electrolysis",
+        bus0=nodes.index,
+        bus1=(nodes.country + f" {suffix}").values,
         p_nom_extendable=True,
         carrier="H2 Electrolysis",
         efficiency=costs.at["electrolysis", "efficiency"],
@@ -2384,14 +2408,27 @@ def add_h2_production_tyndp(n, nodes, buses_h2_z1, costs, options={}):
         lifetime=costs.at["electrolysis", "lifetime"],
     )
 
+    # Add electorlysis to Z2
+    if options["h2_zones_tyndp"]:
+        n.add(
+            "Link",
+            nodes.index + " H2 Z2 Electrolysis",
+            bus0=nodes.index,
+            bus1=(nodes.country + " H2 Z2").values,
+            p_nom_extendable=True,
+            carrier="H2 Electrolysis",
+            efficiency=costs.at["electrolysis", "efficiency"],
+            capital_cost=costs.at["electrolysis", "capital_cost"],
+            lifetime=costs.at["electrolysis", "lifetime"],
+        )
+
     if options["SMR_cc"]:
-        logger.info("Adding Z1 dummy SMR CC.")
         # TODO: this does currently only work for no gas spatial
         n.add(
             "Link",
-            buses_h2_z1 + " SMR CC",
+            buses_h2 + " SMR CC",
             bus0=spatial.gas.nodes,
-            bus1=buses_h2_z1,
+            bus1=buses_h2,
             bus2="co2 atmosphere",
             bus3=spatial.co2.nodes,
             p_nom_extendable=True,
@@ -2404,13 +2441,12 @@ def add_h2_production_tyndp(n, nodes, buses_h2_z1, costs, options={}):
         )
 
     if options["SMR"]:
-        logger.info("Adding Z1 dummy SMR.")
         # TODO: this does currently only work for no gas spatial
         n.add(
             "Link",
-            buses_h2_z1 + " SMR",
+            buses_h2 + " SMR",
             bus0=spatial.gas.nodes,
-            bus1=buses_h2_z1,
+            bus1=buses_h2,
             bus2="co2 atmosphere",
             p_nom_extendable=True,
             carrier="SMR",
@@ -2425,16 +2461,15 @@ def add_h2_production_tyndp(n, nodes, buses_h2_z1, costs, options={}):
         v if isinstance(v, bool) else any(v.values())
         for v in options["methanol"].values()
     ):
-        logger.info("Adding Z1 dummy ATR.")
         add_carrier_buses(
             n=n, carrier="methanol", costs=costs, spatial=spatial, options=options
         )
         # TODO: this does currently only work for no methanol spatial
         n.add(
             "Link",
-            buses_h2_z1 + " ATR",
+            buses_h2 + " ATR",
             bus0=spatial.methanol.nodes,
-            bus1=buses_h2_z1,
+            bus1=buses_h2,
             bus2="co2 atmosphere",
             p_nom_extendable=True,
             carrier="ATR",
@@ -2449,7 +2484,7 @@ def add_h2_production_tyndp(n, nodes, buses_h2_z1, costs, options={}):
 
 def add_h2_dres_tyndp(n, spatial, buses_h2_z2, costs):
     """
-    Adds TYNDP Z2 DRES electricity buses and electrolysers.
+    Adds TYNDP Z2 DRES electricity buses and electrolyzers.
 
     Parameters
     ----------
@@ -2468,7 +2503,7 @@ def add_h2_dres_tyndp(n, spatial, buses_h2_z2, costs):
         The function modifies the network object in-place by adding components.
     """
 
-    logger.info("Adding Z2 dummy DRES electricity buses and electrolysers.")
+    logger.info("Adding Z2 dummy DRES electricity buses and electrolyzers.")
     n.add(
         "Bus",
         buses_h2_z2 + " DRES",
@@ -2493,7 +2528,7 @@ def add_h2_dres_tyndp(n, spatial, buses_h2_z2, costs):
     )
 
 
-def add_h2_reconversion_tyndp(n, spatial, nodes, buses_h2_z2, costs, options=None):
+def add_h2_reconversion_tyndp(n, spatial, nodes, buses_h2, costs, options=None):
     """
     Adds TYNDP H2 reconversion with options for Fuel cells, H2 turbines and methanation.
 
@@ -2505,8 +2540,8 @@ def add_h2_reconversion_tyndp(n, spatial, nodes, buses_h2_z2, costs, options=Non
         Namespace object with spatial nodes for different carriers such as `h2_tyndp`
     nodes : pd.Index
         Pandas Index of electricity node locations/nodes
-    buses_h2_z2 : SimpleNamespace
-        Namespace object with spatial nodes of H2 Z2 buses
+    buses_h2 : pd.Index
+        Pandas Index of hydrogen nodes to which H2 reconversion technologies will connect
     costs : pd.DataFrame
         Technology cost assumptions
     options : dict, optional
@@ -2523,14 +2558,15 @@ def add_h2_reconversion_tyndp(n, spatial, nodes, buses_h2_z2, costs, options=Non
         The function modifies the network object in-place by adding components.
     """
 
+    suffix = "H2 Z2" if options["h2_zones_tyndp"] else "H2"
+
     if options["methanation"]:
-        logger.info("Adding Z2 dummy methanation.")
         # TODO: this does currently only work for no gas spatial and no co2 spatial
         n.add(
             "Link",
-            buses_h2_z2,
+            buses_h2,
             suffix=" Sabatier",
-            bus0=buses_h2_z2,
+            bus0=buses_h2,
             bus1=spatial.gas.nodes,
             bus2=spatial.co2.nodes,
             p_nom_extendable=True,
@@ -2545,12 +2581,10 @@ def add_h2_reconversion_tyndp(n, spatial, nodes, buses_h2_z2, costs, options=Non
         )
 
     if options["hydrogen_fuel_cell"]:
-        logger.info("Adding Z2 hydrogen fuel cell for re-electrification.")
-        # TODO: Improve with TYNDP specific technology and cost assumptions
         n.add(
             "Link",
-            nodes.index + " H2 Z2 h2-fuel-cell",
-            bus0=nodes.country.values + " H2 Z2",
+            nodes.index + f" {suffix} h2-fuel-cell",
+            bus0=(nodes.country + f" {suffix}").values,
             bus1=nodes.index,
             p_nom_extendable=False,
             carrier="h2-fuel-cell",
@@ -2562,13 +2596,12 @@ def add_h2_reconversion_tyndp(n, spatial, nodes, buses_h2_z2, costs, options=Non
 
     if options["hydrogen_turbine"]:
         logger.info(
-            "Adding Z2 hydrogen turbine for re-electrification. Assuming CCGT technology costs."
+            "Adding hydrogen turbine for re-electrification. Assuming CCGT technology costs."
         )
-        # TODO: Improve with TYNDP specific technology and cost assumptions
         n.add(
             "Link",
-            nodes.index + " H2 Z2 h2-ccgt",
-            bus0=nodes.country.values + " H2 Z2",
+            nodes.index + f" {suffix} h2-ccgt",
+            bus0=(nodes.country + f" {suffix}").values,
             bus1=nodes.index,
             p_nom_extendable=False,
             carrier="h2-ccgt",
@@ -2580,7 +2613,7 @@ def add_h2_reconversion_tyndp(n, spatial, nodes, buses_h2_z2, costs, options=Non
         )
 
 
-def add_h2_grid_tyndp(n, nodes, h2_pipes_file, interzonal_file, costs):
+def add_h2_grid_tyndp(n, nodes, h2_pipes_file, interzonal_file, costs, options):
     """
     Adds TYNDP hydrogen pipelines and interzonal (Z1 <-> Z2) connections.
 
@@ -2596,6 +2629,9 @@ def add_h2_grid_tyndp(n, nodes, h2_pipes_file, interzonal_file, costs):
         Path to CSV file containing prepped H2 interzonal connection data
     costs : pd.DataFrame
         Technology cost assumptions
+    options : dict
+        Dictionary of configuration options. Key options include:
+        - h2_zones_tyndp : bool
 
     Returns
     -------
@@ -2603,7 +2639,9 @@ def add_h2_grid_tyndp(n, nodes, h2_pipes_file, interzonal_file, costs):
         The function modifies the network object in-place by adding components.
     """
 
-    h2_pipes = create_h2_topology_tyndp(n=n, fn_h2_network=h2_pipes_file)
+    h2_pipes = create_h2_topology_tyndp(
+        n=n, fn_h2_network=h2_pipes_file, options=options
+    )
     interzonal = pd.read_csv(interzonal_file, index_col=0)
 
     logger.info("Adding TYNDP H2 reference grid pipelines.")
@@ -2701,12 +2739,14 @@ def add_h2_storage_tyndp(
             .loc[:, 0]
         )
 
-        logger.info("Adding dummy TYNDP H2 underground storage for H2 Z2")
+        logger.info("Adding TYNDP H2 underground storage")
+
+        suffix = "H2 Z2" if options["h2_zones_tyndp"] else "H2"
 
         n.add(
             "Store",
-            h2_caverns.index + " H2 Z2 Cavern Store",
-            bus=h2_caverns.index + " H2 Z2",
+            h2_caverns.index + f" {suffix} Cavern Store",
+            bus=h2_caverns.index + f" {suffix}",
             e_nom_extendable=True,
             e_nom_max=h2_caverns.values,
             e_cyclic=True,
@@ -2716,20 +2756,21 @@ def add_h2_storage_tyndp(
         )
 
     # add overground hydrogen tank storage to all H2 Z1 nodes
-    tech = "hydrogen storage tank type 1 including compressor"
+    if not buses_h2_z1.empty:
+        tech = "hydrogen storage tank type 1 including compressor"
 
-    logger.info("Adding dummy TYNDP H2 tank storage for H2 Z1")
+        logger.info("Adding TYNDP H2 tank storage for H2 Z1")
 
-    n.add(
-        "Store",
-        buses_h2_z1 + " Tank Store",
-        bus=buses_h2_z1,
-        e_nom_extendable=True,
-        e_cyclic=True,
-        carrier="H2 Store",
-        capital_cost=costs.at[tech, "capital_cost"],
-        lifetime=costs.at[tech, "lifetime"],
-    )
+        n.add(
+            "Store",
+            buses_h2_z1 + " Tank Store",
+            bus=buses_h2_z1,
+            e_nom_extendable=True,
+            e_cyclic=True,
+            carrier="H2 Store",
+            capital_cost=costs.at[tech, "capital_cost"],
+            lifetime=costs.at[tech, "lifetime"],
+        )
 
 
 def add_h2_topology_tyndp(
@@ -2742,6 +2783,7 @@ def add_h2_topology_tyndp(
     cavern_types,
     costs,
     options,
+    h2_demand_file,
 ):
     """
     Add TYNDP H2 topology to the network.
@@ -2775,6 +2817,8 @@ def add_h2_topology_tyndp(
         Technology cost assumptions
     options : dict, optional
        Dictionary of configuration options. Defaults to empty dict if not provided.
+    h2_demand_file : str
+        Path to CSV file containing exogenous hydrogen demand time series
 
 
     Returns
@@ -2790,21 +2834,13 @@ def add_h2_topology_tyndp(
     # add H2 as carrier
     n.add("Carrier", "H2")
 
-    # filter for electricity nodes and H2 Z1 and Z2 buses
+    # filter for electricity nodes and H2 buses
     nodes = n.buses.loc[pop_layout.index, :].query(
         "country in @spatial.h2_tyndp.country"
     )
-    buses_h2_z2 = spatial.h2_tyndp.nodes[
-        ~spatial.h2_tyndp.nodes.str.contains("IB")
-        & spatial.h2_tyndp.nodes.str.contains("Z2")
-    ]
-    buses_h2_z1 = spatial.h2_tyndp.nodes[
-        ~spatial.h2_tyndp.nodes.str.contains("IB")
-        & spatial.h2_tyndp.nodes.str.contains("Z1")
-    ]
 
-    # add H2 Z1 and Z2 buses
-    logger.info("Adding Z1 and Z2 H2 nodes.")
+    # add H2 Buses
+    logger.info("Adding TYNDP H2 nodes.")
 
     n.add(
         "Bus",
@@ -2817,20 +2853,26 @@ def add_h2_topology_tyndp(
         unit="MWh_LHV",
     )
 
+    # define H2 topology
+    buses_h2_z1 = spatial.buses_h2_z1
+    buses_h2_z2 = spatial.buses_h2_z2
+
     # add H2 production (Z1: Electrolysis, SMR (optional), SMR CC (optional), ATR; Z2: Electrolysis)
+    buses_h2 = buses_h2_z1 if options["h2_zones_tyndp"] else buses_h2_z2
     add_h2_production_tyndp(
-        n=n, nodes=nodes, buses_h2_z1=buses_h2_z1, costs=costs, options=options
+        n=n, nodes=nodes, buses_h2=buses_h2, costs=costs, options=options
     )
 
     # add H2 DRES electricity nodes and Electrolysis to H2 Z2
-    add_h2_dres_tyndp(n=n, spatial=spatial, buses_h2_z2=buses_h2_z2, costs=costs)
+    if options["h2_zones_tyndp"]:
+        add_h2_dres_tyndp(n=n, spatial=spatial, buses_h2_z2=buses_h2_z2, costs=costs)
 
     # add H2 reconversion (Fuel cells (optional), H2 turbines (optional), methanation (optional))
     add_h2_reconversion_tyndp(
         n=n,
         spatial=spatial,
         nodes=nodes,
-        buses_h2_z2=buses_h2_z2,
+        buses_h2=buses_h2_z2,
         costs=costs,
         options=options,
     )
@@ -2842,9 +2884,10 @@ def add_h2_topology_tyndp(
         h2_pipes_file=h2_pipes_file,
         interzonal_file=interzonal_file,
         costs=costs,
+        options=options,
     )
 
-    # add H2 storage (Z1: H2 tanks; Z2: Salt caverns)
+    # add H2 storage (Z1: H2 tanks; Z2/NT H2 nodes: Salt caverns)
     add_h2_storage_tyndp(
         n=n,
         cavern_types=cavern_types,
@@ -2852,6 +2895,44 @@ def add_h2_topology_tyndp(
         buses_h2_z1=buses_h2_z1,
         costs=costs,
         options=options,
+    )
+
+    # add exogenous hydrogen demand
+    add_h2_demand_tyndp(n=n, h2_demand_file=h2_demand_file)
+
+
+def add_h2_demand_tyndp(n, h2_demand_file):
+    """
+    Add exogenous TYNDP hydrogen demand to the network.
+
+    Parameters
+    ----------
+    n : pypsa.Network
+        The PyPSA network container object
+    h2_demand_file : str
+        Path to CSV file containing exogenous hydrogen demand time series
+    """
+    logger.info("Add exogenous hydrogen demand to network")
+
+    demand = pd.read_csv(h2_demand_file, index_col=0, parse_dates=True)
+
+    # check for missing buses
+    h2_buses = n.buses[n.buses.carrier == "H2"].index
+    if not demand.columns.difference(h2_buses).empty:
+        missing = demand.columns.difference(h2_buses)
+        logger.warning(
+            f"The following H2 demand buses are not in the network and are dropped {missing}"
+        )
+        demand.drop(missing, inplace=True, axis=1)
+
+    # add exogenous demand
+    n.add(
+        "Load",
+        demand.columns,
+        suffix=" exogenous demand",
+        bus=demand.columns,
+        carrier="H2 exogenous demand",
+        p_set=demand,
     )
 
 
@@ -3385,6 +3466,8 @@ def add_gas_and_h2_infrastructure(
     gas_input_nodes,
     spatial,
     options,
+    tyndp_scenario,
+    h2_demand_file,
 ):
     """
     Add storage and grid infrastructure to the network for gas and hydrogen.
@@ -3424,6 +3507,10 @@ def add_gas_and_h2_infrastructure(
         - SMR : bool
         - cc_fraction : float
         - methanation : bool
+    tyndp_scenario : str
+        TYNDP scenario name to determine whether to use TYNDP H2 topology
+    h2_demand_file : str
+        Path to CSV file containing exogenous hydrogen demand data
 
     Returns
     -------
@@ -3454,6 +3541,7 @@ def add_gas_and_h2_infrastructure(
             cavern_types=cavern_types,
             costs=costs,
             options=options,
+            h2_demand_file=h2_demand_file,
         )
     else:
         # add base h2 technologies (carrier, production, reconversion, storage)
@@ -6224,6 +6312,7 @@ def add_industry(
     spatial: SimpleNamespace,
     cf_industry: dict,
     investment_year: int,
+    tyndp_scenario: str,
 ):
     """
     Add industry and their corresponding carrier buses to the network.
@@ -6256,6 +6345,8 @@ def add_industry(
         Industry-specific configuration parameters
     investment_year : int
         Year for which investment costs should be considered
+    tyndp_scenario : str
+        TYNDP scenario name for scenario-specific adjustments
     HeatSystem : Enum
         Enumeration defining different heat system types
 
@@ -6415,30 +6506,11 @@ def add_industry(
     )
 
     if options["h2_topology_tyndp"]:
-        # TODO Link properly Industry to H2 topology
-        nodes_ind_h2 = spatial.h2_tyndp.nodes[spatial.h2_tyndp.nodes.str.contains("Z2")]
-        nodes_ind_h2 = nodes_ind_h2[~(nodes_ind_h2.isin(["IBFI H2 Z2", "IBIT H2 Z2"]))]
-        nodes_ind = n.buses.loc[nodes_ind_h2].country.values + "00"
-        nodes_ind[nodes_ind == "IT00"] = "ITCN"
-        nodes_ind[nodes_ind == "DK00"] = "DKE1"
-        nodes_ind[nodes_ind == "LU00"] = "LUG1"
-        nodes_ind[nodes_ind == "NO00"] = "NOS0"
-        nodes_ind[nodes_ind == "SE00"] = "SE01"
+        suffix = "H2 Z2" if options["h2_zones_tyndp"] else "H2"
+        nodes_ind_h2 = pd.Index(pop_layout.ct + f" {suffix}")
+
     else:
-        nodes_ind = nodes
         nodes_ind_h2 = nodes + " H2"
-    industrial_demand_zones = industrial_demand.reindex(nodes_ind, fill_value=0)
-    n.add(
-        "Load",
-        nodes_ind,  # TODO Improve assumptions
-        suffix=" H2 Z2 for industry"
-        if options["h2_topology_tyndp"]
-        else " H2 for industry",  # TODO Improve assumptions
-        bus=nodes_ind_h2,  # TODO Improve assumptions
-        carrier="H2 for industry",
-        p_set=industrial_demand_zones["hydrogen"]
-        / nhours,  # TODO Improve assumptions for zones
-    )
 
     # methanol for industry
 
@@ -6481,10 +6553,10 @@ def add_industry(
     # TODO Link properly Industry to H2 topology
     n.add(
         "Link",
-        nodes_ind + " methanolisation",  # TODO Improve this assumption
+        nodes + " methanolisation",  # TODO Improve this assumption
         bus0=nodes_ind_h2,  # TODO Improve this assumption
         bus1=spatial.methanol.nodes,
-        bus2=nodes_ind,  # TODO Improve this assumption
+        bus2=nodes,
         bus3=spatial.co2.nodes,
         carrier="methanolisation",
         p_nom_extendable=True,
@@ -6522,10 +6594,12 @@ def add_industry(
                 )
 
     # TODO Link properly Industry to H2 topology
+    buses_h2 = spatial.buses_h2_z2 if options["h2_topology_tyndp"] else nodes_ind_h2
+    ft_name = buses_h2.str.replace(" H2 Z2", "").str.replace(" H2", "")
     n.add(
         "Link",
-        nodes_ind + " Fischer-Tropsch",  # TODO Improve assumptions
-        bus0=nodes_ind_h2,  # TODO Improve assumptions
+        ft_name + " Fischer-Tropsch",  # TODO Improve assumptions
+        bus0=buses_h2,  # TODO Improve assumptions
         bus1=spatial.oil.nodes,
         bus2=spatial.co2.nodes,
         carrier="Fischer-Tropsch",
@@ -7904,6 +7978,7 @@ def add_import_options(
     options: dict,
     gas_input_nodes: pd.DataFrame,
     h2_imports_tyndp_fn: str,
+    tyndp_scenario: str,
 ):
     """
     Add green energy import options.
@@ -7918,6 +7993,8 @@ def add_import_options(
         Locations of gas input nodes split by LNG and pipeline.
     h2_imports_tyndp_fn: str,
         Path to file containing H2 import potentials, maximum capacity, offer quantity and marginal cost from TYNDP input data
+    tyndp_scenario : str
+        TYNDP scenario name to be used for H2 imports.
     """
 
     import_config = options["imports"]
@@ -8036,11 +8113,12 @@ def add_import_options(
                 marginal_cost=import_potentials_h2.marginal_cost.values,
                 e_sum_max=import_potentials_h2.e_sum_max.values,
             )
+            suffix = "H2 Z2" if options["h2_zones_tyndp"] else "H2"
             n.add(
                 "Link",
                 import_potentials_h2.index,
                 bus0=import_potentials_h2.Corridor.values + " H2 import",
-                bus1=import_potentials_h2.bus1.values + " H2 Z2",
+                bus1=import_potentials_h2.bus1.values + f" {suffix}",
                 p_nom_extendable=False,
                 p_nom=import_potentials_h2.p_nom.values,
                 bidirectional=False,
@@ -8079,6 +8157,7 @@ if __name__ == "__main__":
 
     options = snakemake.params.sector
     cf_industry = snakemake.params.industry
+    tyndp_scenario = snakemake.params.tyndp_scenario
 
     investment_year = int(snakemake.wildcards.planning_horizons)
 
@@ -8164,6 +8243,7 @@ if __name__ == "__main__":
         options,
         offshore_buses_fn=buses_oh_file,
         buses_h2_file=buses_h2_file,
+        tyndp_scenario=tyndp_scenario,
     )
 
     if snakemake.params.foresight in ["overnight", "myopic", "perfect"]:
@@ -8258,6 +8338,8 @@ if __name__ == "__main__":
         gas_input_nodes=gas_input_nodes,
         spatial=spatial,
         options=options,
+        tyndp_scenario=tyndp_scenario,
+        h2_demand_file=snakemake.input.h2_demand,
     )
 
     if enable_pemmdb_caps:
@@ -8401,6 +8483,7 @@ if __name__ == "__main__":
             spatial=spatial,
             cf_industry=cf_industry,
             investment_year=investment_year,
+            tyndp_scenario=tyndp_scenario,
         )
 
     if options["shipping"]:
@@ -8530,6 +8613,7 @@ if __name__ == "__main__":
             options=options,
             gas_input_nodes=gas_input_nodes,
             h2_imports_tyndp_fn=snakemake.input.h2_imports_tyndp,
+            tyndp_scenario=tyndp_scenario,
         )
 
     if options["gas_distribution_grid"]: